feat(like.controller): validate videoId and check video existence in toggleVideoLike

+++ backend/src/controller/like.controller.js
@@ -2,9 +2,21 @@ import { asyncHandler } from "../util/asyncHandler.js";
 import { Like } from "../model/like.model.js";
 import { ApiResponse } from "../util/ApiResponse.js";
 import { ApiError } from "../util/ApiError.js";
+import { isValidObjectId } from "mongoose";
+import { Video } from "../model/video.model.js";
 
 const toggleVideoLike = asyncHandler(async (req, res) => {
     const { videoId } = req.params;
+
+    if(!videoId || !isValidObjectId(videoId)){
+        throw new ApiError(400, "invalid videoId")
+    }
+
+    const isVideoExists = await Video.exists({_id: videoId})
+
+    if(!isVideoExists){
+        throw new ApiError(404, "video not found")
+    }
 
     const deletedLike = await Like.findOneAndDelete({
         likedBy: req.user?._id,
